requirements.txt
absl-py==1.4.0
accelerate==0.25.0
adan-pytorch==0.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.4.1
alembic==1.10.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
av==10.0.0
axial-positional-embedding==0.2.1
bleach==6.1.0
braindecode==0.8.1
cachetools==5.3.1
cffi==1.16.0
cftime==1.6.3
charset-normalizer==3.1.0
cmaes==0.9.1
cmake==3.26.1
colorama==0.4.6
colorlog==6.7.0
CoLT5-attention==0.10.20
configparser==6.0.0
contourpy==1.0.7
coverage==7.3.2
cramjam==2.8.1
cycler==0.11.0
decord==0.6.0
dill==0.3.8
docstring-inheritance==2.0.2
easydev==0.12.1
EDFlib-Python==1.0.8
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastparquet==2024.2.0
filelock==3.11.0
fire==0.5.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
fvcore==0.1.5.post20221221
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.0
h5py==3.8.0
hjson==3.1.0
huggingface-hub==0.19.4
idna==3.4
imageio==2.28.1
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.2.0
kaggle==1.6.6
kiwisolver==1.4.4
kymatio==0.3.0
lazy_loader==0.3
librosa==0.10.1
lightning-utilities==0.8.0
linear-attention-transformer==0.19.1
linformer==0.2.3
lion-pytorch==0.1.2
lit==16.0.0
llvmlite==0.42.0
lmdb==1.4.1
local-attention==1.9.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
medmnist==2.2.2
memory-profiler==0.61.0
mne==1.4.2
mne-bids==0.13
mne-connectivity==0.5.0
moabb==1.0.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
netCDF4==1.6.5
networkx==3.1
ninja==1.11.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
#nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
#nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
optuna==3.1.1
pandarallel==1.6.5
pandas==1.5.3
parse==1.19.1
patsy==0.5.6
peft==0.7.0
Pillow==9.5.0
piq==0.7.0
plotly==5.17.0
pluggy==1.3.0
pooch==1.7.0
portalocker==2.7.0
product-key-memory==0.2.10
protobuf==4.23.4
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pyhealth==1.1.4
pyparsing==3.0.9
pyriemann==0.5
pytest==7.4.3
python-slugify==8.0.4
pytorch-lightning==2.1.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rdkit==2023.9.5
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.10.1
seaborn==0.12.2
skorch==0.15.0
soundfile==0.12.1
soxr==0.3.7
spectrum==0.8.1
SQLAlchemy==2.0.9
statsmodels==0.14.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.2.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.16
tokenizers==0.15.0
tomli==2.0.1
#torch==2.0.0
torch-lfilter==0.0.3
#torchaudio==2.0.1
torchcam==0.4.0
torcheeg==1.1.0
torchinfo==1.8.0
torchlibrosa==0.1.0
torchmetrics==0.11.4
torchsampler==0.1.2
torchsummary==1.5.1
#torchvision==0.15.1
tqdm==4.65.0
transformers==4.35.2
#triton==2.0.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.3.6
xarray==2023.10.1
xmltodict==0.13.0
yacs==0.1.8
yarl==1.9.2
